Add files via upload
This is a python file to solve 8 Piece puzzle Solver using BFS(Brute Force Search). This is not efficient way to solve the problem as it is computationally expensive. It explore all the nodes and back-track its way to find solution.

It is bound to find a way for any possible* node in lower bound of time around 20 minutes.